refactor: Notice dto 네이밍 및 디렉토리 구조 정리
