+++ src/main/java/com/example/Tokkit_server/dto/notice/NoticeRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.Tokkit_server.dto.notices;
+package com.example.Tokkit_server.dto.notice;
 
 import com.example.Tokkit_server.domain.Notice;
 import lombok.Getter;
@@ -10,7 +10,7 @@ public class NoticeRequestDto {
     private String title;
     private String content;
 
-    public Notice toEntity() {
+    public Notice to() {
         return Notice.builder()
                 .title(title)
                 .content(content)
