+++ src/main/java/com/example/Tokkit_server/service/query/NoticeQueryService.java
@@ -3,7 +3,7 @@ package com.example.Tokkit_server.service.query;
 import com.example.Tokkit_server.apiPayload.code.status.ErrorStatus;
 import com.example.Tokkit_server.apiPayload.exception.GeneralException;
 import com.example.Tokkit_server.domain.Notice;
-import com.example.Tokkit_server.dto.notices.NoticeResponseDto;
+import com.example.Tokkit_server.dto.notice.NoticeResponseDto;
 import com.example.Tokkit_server.repository.NoticeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -18,13 +18,13 @@ public class NoticeQueryService {
 
     public Page<NoticeResponseDto> getNotices(int page) {
         Page<Notice> notices = noticeRepository.findAllByIsDeletedFalseOrderByCreatedAtDesc(PageRequest.of(page, SIZE));
-        return notices.map(NoticeResponseDto::fromEntity);
+        return notices.map(NoticeResponseDto::from);
     }
 
     public NoticeResponseDto getNotice(Long noticeId) {
         Notice notice = noticeRepository.findById(noticeId)
                 .orElseThrow(() -> new GeneralException(ErrorStatus.NOTICE_NOT_FOUND));
 
-        return NoticeResponseDto.fromEntity(notice);
+        return NoticeResponseDto.from(notice);
     }
 }
